+++ packages/site/src/componentList.ts
@@ -65,6 +65,21 @@ export const componentList = [
     imageURL: "/Countdown.png",
   },
   {
+    title: "DataDump",
+    to: "/components/DataDump",
+    imageURL: "/DataDump.png",
+  },
+  {
+    title: "DataList",
+    to: "/components/DataList",
+    imageURL: "/DataList.png",
+  },
+  {
+    title: "DatePicker",
+    to: "/components/DatePicker",
+    imageURL: "/DatePicker.png",
+  },
+  {
     title: "Disclosure",
     to: "/components/Disclosure",
     imageURL: "/Disclosure.png",
@@ -81,6 +96,32 @@ export const componentList = [
     to: "/components/Emphasis",
     imageURL: "/Emphasis.png",
     additionalMatches: ["Highlight", "Strong", "Bold", "Italic"],
+  },
+  {
+    title: "FormatDate",
+    to: "/components/FormatDate",
+    imageURL: "/FormatDate.png",
+  },
+  {
+    title: "FormatEmail",
+    to: "/components/FormatEmail",
+    imageURL: "/FormatEmail.png",
+  },
+  {
+    title: "FormatRelativeDateTime",
+    to: "/components/FormatRelativeDateTime",
+    imageURL: "/FormatRelativeDateTime.png",
+  },
+  {
+    title: "FormatTime",
+    to: "/components/FormatTime",
+    imageURL: "/FormatTime.png",
+  },
+  {
+    title: "Glimmer",
+    to: "/components/Glimmer",
+    imageURL: "/Glimmer.png",
+    additionalMatches: ["Skeleton", "Shimmer"],
   },
   {
     title: "Heading",
@@ -108,6 +149,12 @@ export const componentList = [
     ],
   },
   {
+    title: "InputDate",
+    to: "/components/InputDate",
+    imageURL: "/InputDate.png",
+    additionalMatches: ["Datepicker", "Datetime Picker", "Calendar"],
+  },
+  {
     title: "Link",
     to: "/components/Link",
     imageURL: "/Link.png",
@@ -124,6 +171,11 @@ export const componentList = [
     to: "/components/Spinner",
     imageURL: "/Spinner.png",
     additionalMatches: ["Loader", "Loading", "Twirly", "Wait", "Busy"],
+  },
+  {
+    title: "StatusIndicator",
+    to: "/components/StatusIndicator",
+    imageURL: "/StatusIndicator.png",
   },
   {
     title: "StatusLabel",